-viewModel for labs index -number of available computers in lab
