+++ CAMS/CAMS/CAMS.csproj
@@ -222,7 +222,6 @@
     <Compile Include="Models\ComputerLab.cs">
       <DependentUpon>CAMS_DatabaseModel.tt</DependentUpon>
     </Compile>
-    <Compile Include="Models\LabModel.cs" />
     <Compile Include="Models\Department.cs">
       <DependentUpon>CAMS_DatabaseModel.tt</DependentUpon>
     </Compile>
@@ -230,6 +229,7 @@
     <Compile Include="Models\Lab.cs">
       <DependentUpon>CAMS_DatabaseModel.tt</DependentUpon>
     </Compile>
+    <Compile Include="Models\LabsViewModel.cs" />
     <Compile Include="Models\LdapAuthentication.cs" />
     <Compile Include="Models\ManageViewModels.cs" />
     <Compile Include="Models\User.cs">
