+++ CAMS/CAMS/Views/Labs/Details.cshtml
@@ -1,4 +1,4 @@
-﻿@model CAMS.Models.LabModel
+﻿@model CAMS.Models.LabDetailsViewModel
 
 
 @{
